Move permission tests to their own file
